added incomplete solution to problem 453
The algorithm is wrong. It works by counting the number of ways to choose 4 points such that at least 3 are collinear and subtracting that from the number of total ways to choose 4 points. But this is wrong because a set of four points does NOT uniquely identify a quadrilateral (it can correspond to multiple, depending on how you connect the points).