solved problem 61
This problem confused me for a while even after I solved it. I think this is because of the weird wording in the problem statement: apparently they consider the phrase "each polygonal type is represented by a different number in the set" to allow some numbers to have multiple polygonal types. In particular, apparently 8128 is allowed in the solution set even though it is both triangular and hexagonal.

I guess this does kind of make sense when you think about it since hexagonals are "represented" by 8128 in the solution and triangulars are "represented" by 8256, and those are different numbers.  But I still blame the wording of the problem.