started adding method PartialCutting.equals
After writing the tests, I realized this is probably not a good idea, since I want the method `allCuttings` to  output the cuttings in their canonical form anyways (i.e. with line edge's vertices listed in increasing order, with edges listed in order of their lowest vertex, with regions listed in order of their first edge in the edge list, and with region's edges listed clockwise starting from the lowest vertex).