fixed numRearrangements + removed incorrect test
The removed test was testing for behavior that I am no longer requiring -- i.e. if there are subtrees that are rearrangements of each other but are not deeply equal, then I am okay with the method producing the wrong answer. This is because when solving the actual problem, all the trees this algorithm is run on will be in their 'standard form' and so the case where it gives the incorrect answer will never occur.