fixed incorrect test cases
See page 89 of notebook 4 for a drawing of the test case.

The first incorrect test case was because when I was calculating the answer by hand, I missed one of the possible types of cuttings. The second test case was caused because I miscounted the number of rotational symmetries of one of the types.