optimized: sorted probabilities in descending order
This is an optimization because it allows the conditions on lines 70 and 75 to prune branches more often.

After implementing this optimization, I ran it (on the actual problem) and it completed 25 of 49 steps (which is much less good than it appears, since each step takes around twice as long as the previous step). It got this far because it was able to prune 97% of the nodes (where the 97% number includes descendants of pruned nodes as well).