added code to log upper and lower bounds
The idea here is that if the upper and lower bounds ever get close enough together, I can stop the computation halfway through, since Project Euler only requires 10 digits of precision.

This was not particularly successful: after running it for around 1 minute(?), the lower bound was around 0.58, while the upper bound was still very close to 1. (The lower bound seemed to be increasing at a steady rate with each iteration, but of course the iterations get really slow as you go along).