refactored: removed upperBound variable
Proof of correctness: since it iterates over sets S in increasing order of max(sum(S), product(S)), all the sets for which sum(S) = product(S) will be found in increasing order of the sum/product. Because of this, once it has found a product-sum number for each set size, those product-sum numbers are necessarily minimal, and so no further searching is required.