optimized: removed `setSizeParity` and `isEmpty`
This is an optimization because fewer arguments will cause it to lookup cached values more often since it's memoized.

Performance tests: I ran it on the same input as the previous commit, and it took 3.441s, 3.256s, and 3.293s -- an improvement over the previous times of 5.910s, 5.731s, and 5.841s.