fixed the bug!
The issue was that after changing the array to a map, the iteration order was different (because maps iterate over key/value pairs in order of insertion). The solution was to keep a priority queue of all the numbers that have been factorized so far, then iterate over that in order when we want to iterate over the factorized numbers.

Performance notes: on an input of 2000, the new algorithm took 14.467s and 14.116s, while the old algorithm took 28.061s and 26.927 s.