optimized: used more efficient VectorSet data structure
Performance notes: I ran it on an input of 100,000 again and this time it only took 756ms, 700ms, and 781ms (approximately 10x better than before!).

I also ran it on the larger input of 500,000, and it took 8.485s, 8.293s, and 8.504s.

I also ran it on the input for the problem (10,000,000), and it solved it incorrectly in 28 minutes, producing the incorrect answer of 689680414.