changed some parts to use bigints
I changed the two parts that seemed like they would have the largest intermediate values in the computation (even though I calculated that bigints were not necessary in any part of the program). And indeed, I was correct: changing it to use bigints did not change the answer (although it did make it much slower to compute!).