added an algorithm for problem 103 (not working)
It loops forever on inputs greater than 2. This is because it uses depth-first search, and so if it goes down the wrong branch, it might not find any special sets, and it'll keep searching forever before going back up to go down the correct branch.

I don't see any way to fix this apart from writing a completely different algorithm.