added some code to solve an abstract algebra problem
The problem came up when I was trying to prove that the commutator subgroup is normal. I needed to find a way of writing any element of the form `g x^-1 y^-1 xy g^-1` in the form `a^-1 b^-1 ab` for some a, b in the group. That is what the code computes.