changed custom equality function
Previously there was a function `areEqual` that could be overridden if a group needed custom equality. The problem with this is that it makes checking if a set includes an element very annoying (you have to check each one, even if the equality function is still ===, since there's no way to determine that).