refactored old algorithm
It now counts symmetrical and asymmetrical sculptures separately, so it will now be easier to make it generate a list of sculptures instead of just counting how many there are.

Note: I also removed an optimization (special case for when there's 1 block left), so this will need to be added back if it gets slow again.